revert

+++ frontend/src/services/admin.js
@@ -1,9 +1,8 @@
-import endpoint from "./api";
-const api = endpoint("/admin");
+import api from "./api";
 
 export const admin = {
     registerUser: async (name, email, password, role) => {
-        const response = await api.post(`/register`, {
+        const response = await api.post('/admin/register', {
             name,
             email,
             password,
@@ -13,32 +12,33 @@ export const admin = {
     },
 
     getAllUsers: async () => {
-        const response = await api.get(`/getall`);
+        const response = await api.get('/admin/getall');
         return response.data;
     },
 
     getAllCourses: async () => {
-        const response = await api.get(`/courses`);
+        const response = await api.get('/admin/courses');
         return response.data;
     },
 
     deleteUser: async (userId) => {
-        const response = await api.delete(`/delete-user/${userId}`);
+        const response = await api.delete(`/admin/delete-user/${userId}`);
         return response.data;
     },
 
     updateUser: async (userId, userData) => {
-        const response = await api.put(`/update-user/${userId}`, userData);
+        const response = await api.put(`/admin/update-user/${userId}`, userData);
         return response.data;
     },
 
     fetchAllVideos: async () => {
-        const response = await api.get(`/get-video`);
+        const response = await api.get('/admin/get-video');
         return response.data;
     },
 
     deleteVideo: async (videoId) => {
-        const response = await api.delete(`/delete-video/${videoId}`);
+        const response = await api.delete(`/admin/delete-video/${videoId}`);
         return response.data;
     }
+
 };
